pyproject.toml
[project]
name = "multiagent-poc-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "asyncclick>=8.1.8",
    "azure-ai-agents>=1.0.1",
    "azure-ai-projects>=1.0.0b12",
    "azure-identity>=1.23.0",
    "bs4>=0.0.2",
    "click>=8.2.1",
    "fastapi[standard]>=0.115.14",
    "feedparser>=6.0.11",
    "gradio>=5.35.0",
    "jwcrypto>=1.5.6",
    "mcp[cli]>=1.10.1",
    "openai-agents>=0.1.0",
    "semantic-kernel>=1.32.2",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "uvicorn>=0.35.0",
]
